pyproject.toml
[project]
name = "tenentviz"
version = "0.1.0"
description = ""
authors = [
    {name = "MaximeDemougin",email = "demougin.maxime1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.52.2,<2.0.0)",
    "extra-streamlit-components (>=0.1.81,<0.2.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
